-refactors networkmodels
