+++ app/src/main/java/com/andrew/tasky/agenda/data/networkmodels/AgendaResponse.kt
@@ -1,0 +1,7 @@
+package com.andrew.tasky.agenda.data.networkmodels
+
+data class AgendaResponse(
+    val events: List<EventDto>,
+    val tasks: List<TaskDto>,
+    val reminders: List<ReminderDto>
+)
