+++ app/src/main/java/com/andrew/tasky/agenda/data/networkmodels/AttendeeDto.kt
@@ -1,4 +1,4 @@
-package com.andrew.tasky.agenda.data.networkModels
+package com.andrew.tasky.agenda.data.networkmodels
 
 data class AttendeeDto(
     val email: String,
