+++ app/src/main/java/com/andrew/tasky/agenda/data/networkmodels/CreateEventRequest.kt
@@ -1,4 +1,4 @@
-package com.andrew.tasky.agenda.data.networkModels
+package com.andrew.tasky.agenda.data.networkmodels
 
 import kotlinx.serialization.Serializable
