+++ app/src/main/java/com/andrew/tasky/agenda/data/reminder/ReminderApi.kt
@@ -1,6 +1,6 @@
 package com.andrew.tasky.agenda.data.reminder
 
-import com.andrew.tasky.agenda.data.networkModels.ReminderDto
+import com.andrew.tasky.agenda.data.networkmodels.ReminderDto
 import retrofit2.Response
 import retrofit2.http.*
