+++ app/src/main/java/com/andrew/tasky/agenda/data/task/TaskApi.kt
@@ -1,6 +1,6 @@
 package com.andrew.tasky.agenda.data.task
 
-import com.andrew.tasky.agenda.data.networkModels.TaskDto
+import com.andrew.tasky.agenda.data.networkmodels.TaskDto
 import retrofit2.Response
 import retrofit2.http.*
