+++ app/src/main/java/com/andrew/tasky/auth/data/AuthApi.kt
@@ -1,8 +1,8 @@
 package com.andrew.tasky.auth.data
 
-import com.andrew.tasky.auth.data.networkModels.LoginRequest
-import com.andrew.tasky.auth.data.networkModels.RegisterRequest
-import com.andrew.tasky.auth.data.networkModels.TokenResponse
+import com.andrew.tasky.auth.data.networkmodels.LoginRequest
+import com.andrew.tasky.auth.data.networkmodels.RegisterRequest
+import com.andrew.tasky.auth.data.networkmodels.TokenResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
