+++ app/src/main/java/com/andrew/tasky/auth/data/networkmodels/TokenResponse.kt
@@ -1,4 +1,4 @@
-package com.andrew.tasky.auth.data.networkModels
+package com.andrew.tasky.auth.data.networkmodels
 
 data class TokenResponse(
     val token: String,
