+++ app/src/main/java/com/andrew/tasky/auth/util/GetResourceResult.kt
@@ -2,8 +2,8 @@ package com.andrew.tasky.auth.util
 
 import com.andrew.tasky.R
 import com.andrew.tasky.core.UiText
-import com.andrew.tasky.core.data.ErrorMessageDto
 import com.andrew.tasky.core.data.Resource
+import com.andrew.tasky.core.data.networkmodels.ErrorMessageDto
 import com.andrew.tasky.core.data.util.ApiErrorType
 import com.squareup.moshi.Moshi
 import kotlinx.coroutines.Dispatchers
