+++ app/src/main/java/com/andrew/tasky/core/data/networkmodels/ErrorMessageDto.kt
@@ -1,4 +1,4 @@
-package com.andrew.tasky.core.data
+package com.andrew.tasky.core.data.networkmodels
 
 data class ErrorMessageDto(
     val message: String
